Dirty PoC for lock-less UdpMuxNewAddr
The general idea is this:

We use `flume` channels between the actual `UdpMuxNewAddr` instance
and a set of "handles", one per async-trait that we need to satisfy.

There is only one instance of `UdpMuxNewAddr`, meaning we can access
it via `&mut self` and drop all locks within it.

In the poll function, we can then decide, which events to prioritise
over others. Most importantly, all channels between handles and the
actual instance are "rendezvous channels", meaning, sending will block
until we actually read the item out of the channel. This allows us
to enforce backpressure from the poll function all the way to all
handles and into the task that is interacting with the handle.