Simplify send_buffer handling
Instead of buffering multiple items, we only ever buffer one. To
work off this queue as quickly as possible, we put it at the top
of the loop which allows us to reduce some code duplication by
writing directly to the buffer in case we popped an item off the
queue and going back to the start of the loop which will trigger
a different code branch.